requirements.txt
# Core
Django>=4.2.27,<6.0
python-dotenv>=1.0
python-decouple>=3.8
sqlparse>=0.4.4
pytz

# REST
djangorestframework>=3.15.2,<3.16
django-filter>=24.2
django-cors-headers>=4.4,<5.0
djangorestframework-simplejwt>=5.3

# Translations
django-parler>=2.3
django-parler-rest

# Images / HTML
Pillow>=10.3
bleach>=6.1
xhtml2pdf

# Static files
whitenoise>=6.7
django-compressor

# Forms & Admin
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
django-jazzmin>=2.6.0
django-widget-tweaks>=1.5.0

# Database
psycopg2>=2.9
dj-database-url>=2.2

# Redis / Celery
redis>=5.0,<6.0
celery>=5.4
django-celery-beat
channels[daphne]>=4.1
channels-redis>=4.2.0

# Google Cloud (TRANSLATION OFFICIELLE)
google-cloud-translate>=3.15,<4.0
google-api-core>=2.17,<3.0
google-auth>=2.28,<3.0

# Utils
requests>=2.31
six>=1.16
humanize
qrcode>=7.4
pycryptodome>=3.20

# Communication
twilio
django-anymail[sendgrid,sparkpost]

# Geo
geopy
geoip2
django-ipware
timezonefinder
user-agents

# Payments
stripe

# Firebase
firebase-admin
django-extensions 
graphviz

